deleting unused components and code
